apps/phase2-crewai/pyproject.toml
[project]
name = "phase2-crewai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[google-genai]>=1.12.2",
    "crewai-tools>=1.12.2",
    "python-dotenv>=1.1.1",
    "shared-prompts",
    "tavily-python>=0.7.23",
]

[tool.uv.sources]
shared-prompts = { path = "../../packages/shared-prompts" }
